requirements.txt
asgiref==3.2.7
Django==2.2.10
Pillow==7.1.0
pytz==2019.3
sqlparse==0.3.1
